Simplify repository structure (remove `addon/resolvers/classic`)
The `resolvers` subdirectory was created in order to support both the
classic resolver and the module unification resolver. Some time later,
the MU resolver was removed.

This cleans the repo up a bit to make it much simpler (and more "normal").
