--- addon/addon/resolver.js
@@ -1,1 +1,0 @@
-export { default } from './resolvers/classic';
